+++ Farming Frenzy/Assets/Code/Scripts/Plants/Powers/SpikyPower.cs
@@ -2,6 +2,6 @@ namespace Code.Scripts.Plants.Powers
 {
     public abstract class SpikyPower : Power
     {
-        public abstract int Damage { get; }
+        public abstract float Damage { get; }
     }
 }
